Colorscheme group inheritance
With this commit, the syntax files can define groups that are subsets of
other groups, for example constant.string. This is so that colorschemes
can be more accurate, possibly highlighting strings differently than
numbers for example.

See #176. This doesn't fully close that issue yet because the string
group still needs to be added to all strings in the syntax files.