Change default colorscheme to monokai
The monokai-tc colorscheme has been renamed to plain monokai and has
been made the default colorscheme. It is still a true color colorscheme,
but the automatic 256 color approximation is very good so it can be used
just fine in terminals which do not support true color.

If your terminal only supports 16 colors, used the 'default'
colorscheme.